Correct expected naming convention of log files.
Primo log file names always include the day in the title as a 2
digit number.  Therefore the 10th is "10" and the 2nd is "02.
Previous code expected the day in the title of log files to be
single digit if the day was between 1 and 9 (i.e.
localhost_access_log.2016-02-01.log). If we want this to just
work out-of-the-box we should correct this.